# codex-auth-openapi.yaml
openapi: 3.1.0
info:
  title: Codex OAuth Upstream Endpoint Surface
  version: 0.1.0
  summary: Upstream endpoints observed with Codex OAuth bearer tokens.
  description: |
    This document describes upstream ChatGPT, Codex, and public OpenAI
    endpoints that Guidewire has evidence for with Codex OAuth bearer tokens.
    It intentionally excludes local Guidewire loopback routes.

    Operations use Guidewire vendor extensions to distinguish proven support
    and auth-accepted body validation failures where applicable.
  x-guidewire-scope: upstream-codex-oauth-only
  x-guidewire-primary-evidence:
    - docs/CODEX-TO-RESPONSES-ADAPTER.md
    - docs/CHATGPT-TO-RESPONSES-ADAPTER.md
    - docs/gpt-web-evidence/atlas_voice_re_20260515/INDEX.md
    - fixtures/codex/models-upstream.json
servers:
  - url: https://chatgpt.com
    description: ChatGPT and Codex HTTPS backend surfaces.
  - url: https://api.openai.com
    description: Public OpenAI HTTPS surfaces probed with Codex OAuth.
  - url: wss://chatgpt.com
    description: ChatGPT and Codex WebSocket surfaces.
  - url: wss://api.openai.com
    description: Public OpenAI WebSocket realtime surface.
security:
  - codexOAuthBearer: []
...
paths:
...
  /backend-api/codex/responses:
    post:
      operationId: createCodexResponse
      tags:
        - Codex responses
      summary: Create a Codex backend response over HTTPS/SSE.
      description: |
        Creates a Codex backend response using a flat Responses-shaped request.
        Live validation required `instructions`, list-shaped `input`, and
        `stream: true`. The backend returned `store: false` and streamed
        `response.*` events, including encrypted reasoning output when requested.
      servers:
        - url: https://chatgpt.com
      security:
        - codexOAuthBearer: []
      parameters:
        - $ref: '#/components/parameters/ChatGPTAccountId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CodexResponseCreateRequest'
            examples:
              text:
                value:
                  model: gpt-5.5
                  instructions: You are a helpful assistant.
                  input:
                    - role: user
                      content:
                        - type: input_text
                          text: Reply with OK.
                  stream: true
                  store: false
                  reasoning:
                    effort: medium
                    summary: auto
                  include:
                    - reasoning.encrypted_content
              compactionTrigger:
                value:
                  model: gpt-5.5
                  instructions: You are a helpful assistant.
                  input:
                    - type: compaction_trigger
                  stream: true
                  store: false
      responses:
        '200':
          description: SSE stream containing Responses-style events.
          content:
            text/event-stream:
              schema:
                type: string
              examples:
                textDelta:
                  value: |
                    data: {"type":"response.output_text.delta","delta":"OK"}

                    data: {"type":"response.completed","response":{"store":false}}
            application/json:
              schema:
                $ref: '#/components/schemas/CodexResponseObject'
        '400':
          description: Invalid request shape, including unsupported compaction item aliases.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenAIStyleErrorEnvelope'
      x-guidewire-disposition: proven
      x-guidewire-auth-scope-result: accepted
      x-guidewire-evidence:
        - docs/CODEX-TO-RESPONSES-ADAPTER.md#live-validation-evidence
        - sidecar/src/codex/responsesClient.ts
        - sidecar/src/codex/websocketClient.ts
      x-guidewire-notes:
        - The live compaction request item is `compaction_trigger`, not `context_compaction`.
        - Public retrieve semantics for `store: false` responses must be adapter-owned.

    get:
      operationId: connectCodexResponsesWebSocket
      tags:
        - Codex responses
      summary: Connect to Codex Responses WebSocket.
      description: |
        WebSocket connection to the Codex responses backend. Clients send a
        `response.create` envelope after the WebSocket opens and receive JSON
        `response.*` event frames.
      servers:
        - url: wss://chatgpt.com
      security:
        - codexOAuthBearer: []
      parameters:
        - $ref: '#/components/parameters/ChatGPTAccountId'
        - name: OpenAI-Beta
          in: header
          required: true
          description: Required Codex Responses WebSocket beta header.
          schema:
            type: string
            const: responses_websockets=2026-02-06
        - name: originator
          in: header
          required: true
          schema:
            type: string
            example: guidewire
        - name: x-client-request-id
          in: header
          required: true
          schema:
            type: string
            format: uuid
        - name: session_id
          in: header
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '101':
          description: WebSocket upgrade accepted.
        '401':
          description: Missing or invalid Codex OAuth bearer.
        '429':
          description: Rate limited by Codex backend.
      x-guidewire-protocol: websocket
      x-guidewire-websocket-client-message:
        $ref: '#/components/schemas/CodexResponsesWebSocketCreateMessage'
      x-guidewire-websocket-server-message:
        $ref: '#/components/schemas/ResponseStreamEvent'
      x-guidewire-disposition: implemented
      x-guidewire-auth-scope-result: accepted
      x-guidewire-evidence:
        - sidecar/src/codex/responsesClient.ts
        - sidecar/src/codex/websocketClient.ts
  /backend-api/codex/responses/compact:
    post:
      operationId: compactCodexResponseContext
      tags:
        - Codex responses
      summary: Compact Codex response context.
      description: |
        Forwards a compaction request to the Codex backend. Live validation
        returned preserved user messages plus a `compaction_summary` item with
        encrypted content. Adapter storage should canonicalize that item to
        `compaction` before replay.
      servers:
        - url: https://chatgpt.com
      security:
        - codexOAuthBearer: []
      parameters:
        - $ref: '#/components/parameters/ChatGPTAccountId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CodexCompactRequest'
      responses:
        '200':
          description: Compacted message list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodexCompactResponse'
        '400':
          description: Invalid compaction request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenAIStyleErrorEnvelope'
      x-guidewire-disposition: proven
      x-guidewire-auth-scope-result: accepted
      x-guidewire-evidence:
        - docs/CODEX-TO-RESPONSES-ADAPTER.md#live-validation-evidence
...
components:
  securitySchemes:
    codexOAuthBearer:
      type: http
      scheme: bearer
      bearerFormat: Codex OAuth access token
      description: Codex OAuth bearer token from the Codex/ChatGPT auth flow.
  parameters:
    ChatGPTAccountId:
      name: ChatGPT-Account-Id
      in: header
      required: false
      description: Optional ChatGPT account selector header used by Codex/ChatGPT backend surfaces.
      schema:
        type: string
  schemas:
    OpenAIStyleErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            message:
              type: string
            type:
              type: string
            param:
              type:
                - string
                - 'null'
            code:
              type:
                - string
                - 'null'
          additionalProperties: true
      additionalProperties: true
...
    CodexResponseCreateRequest:
      type: object
      required:
        - model
        - instructions
        - input
        - stream
      properties:
        model:
          type: string
          example: gpt-5.5
        instructions:
          type: string
        input:
          type: array
          items:
            oneOf:
              - $ref: '#/components/schemas/ResponseInputMessage'
              - $ref: '#/components/schemas/CompactionTriggerItem'
              - $ref: '#/components/schemas/CompactionItem'
        stream:
          type: boolean
          const: true
        store:
          type: boolean
          default: false
        reasoning:
          $ref: '#/components/schemas/ReasoningConfig'
        include:
          type: array
          items:
            type: string
          example:
            - reasoning.encrypted_content
        tools:
          type: array
          items:
            type: object
            additionalProperties: true
        tool_choice:
          description: Tool choice object or string, if tools are supplied.
        metadata:
          type: object
          additionalProperties:
            type:
              - string
              - number
              - boolean
              - 'null'
      additionalProperties: true
    CodexResponsesWebSocketCreateMessage:
      allOf:
        - $ref: '#/components/schemas/CodexResponseCreateRequest'
        - type: object
          required:
            - type
          properties:
            type:
              type: string
              const: response.create
      x-guidewire-notes:
        - Codex Responses WebSocket is stricter than HTTP for top-level request fields.
        - Live probes on 2026-05-21 accepted `client_metadata`, `prompt_cache_key`, `reasoning`, `include`, and `tool_choice`, but rejected top-level `metadata`, `originator`, `chatMode`, `temperature`, and arbitrary unknown keys with `400 invalid_request_error`.
        - Keep session/thread metadata in `client_metadata` or headers, not top-level `metadata` or `originator`.
    ResponseInputMessage:
      type: object
      required:
        - role
        - content
      properties:
        role:
          type: string
          enum:
            - user
            - assistant
            - system
            - developer
        content:
          type: array
          items:
            oneOf:
              - $ref: '#/components/schemas/InputTextContent'
              - $ref: '#/components/schemas/InputImageContent'
      additionalProperties: true
    InputTextContent:
      type: object
      required:
        - type
        - text
      properties:
        type:
          type: string
          const: input_text
        text:
          type: string
      additionalProperties: true
    InputImageContent:
      type: object
      required:
        - type
      properties:
        type:
          type: string
          const: input_image
        image_url:
          type: string
        detail:
          type: string
          enum:
            - low
            - high
            - auto
            - original
      additionalProperties: true
    CompactionTriggerItem:
      type: object
      required:
        - type
      properties:
        type:
          type: string
          const: compaction_trigger
      additionalProperties: true
    CompactionItem:
      type: object
      required:
        - type
        - encrypted_content
      properties:
        type:
          type: string
          enum:
            - compaction
            - compaction_summary
        encrypted_content:
          type: string
      additionalProperties: true
    ReasoningConfig:
      type: object
      properties:
        effort:
          type: string
          enum:
            - minimal
            - low
            - medium
            - high
            - xhigh
        summary:
          type: string
          enum:
            - auto
            - none
            - concise
            - detailed
      additionalProperties: true
    CodexResponseObject:
      type: object
      properties:
        id:
          type: string
        object:
          type: string
        model:
          type: string
        store:
          type: boolean
        output:
          type: array
          items:
            type: object
            additionalProperties: true
      additionalProperties: true
    ResponseStreamEvent:
      type: object
      required:
        - type
      properties:
        type:
          type: string
          examples:
            - response.created
            - response.output_text.delta
            - response.output_text.done
            - response.output_item.done
            - response.completed
            - response.done
            - response.failed
        response:
          $ref: '#/components/schemas/CodexResponseObject'
        delta:
          type: string
        item:
          type: object
          additionalProperties: true
        error:
          type: object
          additionalProperties: true
      additionalProperties: true
    CodexCompactRequest:
      type: object
      required:
        - model
        - input
      properties:
        model:
          type: string
          example: gpt-5.5
        input:
          type: array
          items:
            oneOf:
              - $ref: '#/components/schemas/ResponseInputMessage'
              - $ref: '#/components/schemas/CompactionItem'
        instructions:
          type: string
        tools:
          type: array
          items:
            type: object
            additionalProperties: true
        reasoning:
          $ref: '#/components/schemas/ReasoningConfig'
        prompt_cache_key:
          type: string
        prompt_cache_retention:
          type: string
      additionalProperties: true
    CodexCompactResponse:
      type: object
      properties:
        input:
          type: array
          items:
            oneOf:
              - $ref: '#/components/schemas/ResponseInputMessage'
              - $ref: '#/components/schemas/CompactionItem'
        output:
          type: array
          items:
            oneOf:
              - $ref: '#/components/schemas/ResponseInputMessage'
              - $ref: '#/components/schemas/CompactionItem'
      additionalProperties: true
...